fix recursion
