+++ sciris/__init__.py
@@ -8,13 +8,13 @@ of ``sc.sc_parallel.parallelize()``.
 
 # Import everything
 from .sc_version   import *
-from .sc_settings  import *
 from .sc_utils     import *
 from .sc_printing  import *
-from .sc_datetime  import *
 from .sc_nested    import *
-from .sc_math      import *
 from .sc_odict     import *
+from .sc_settings  import *
+from .sc_datetime  import *
+from .sc_math      import *
 from .sc_dataframe import *
 from .sc_fileio    import *
 from .sc_profiling import *
